Code cleanup

+++ myob-server/routes/cards.js
@@ -7,8 +7,9 @@ const Favorite = require("../models/Favorite");
 const _ = require("lodash")
 const bcrypt = require("bcryptjs")
 const jwt = require("jsonwebtoken")
+const chalk = require("chalk")
 
-//Get all cards 
+//Get all cards
 router.get("/", async (req, res) => {
     try {
         const cards = await Card.find();
@@ -19,11 +20,10 @@ router.get("/", async (req, res) => {
     }
 })
 
-//Get my-cards
+//Get my-cards 
 router.get("/my-cards", auth, async (req, res) => {
     try {
         const cards = await Card.find({ owner: req.payload.email });
-        console.log(cards);
         if (!cards) return res.status(400).send("There are no Business Cards to show...");
         res.status(200).send(cards);
     } catch (error) {
@@ -42,7 +42,7 @@ router.get("/:_id", async (req, res) => {
     }
 })
 
-//Create Card - by business\admin user 
+//Create Card - business\admin user only
 const cardSchema = joi.object({
     title: joi.string().required().min(2), subtitle: joi.string().required().min(2), description: joi.string().required().min(20),
     phone: joi.string().required().min(2), email: joi.string().required().email(), webSite: joi.string().min(10), businessImage: joi.object({ url: joi.string().min(2), alt: joi.string().min(2) }), address: joi.object({ country: joi.string().required().min(2), state: joi.string().min(0), city: joi.string().required().min(2), street: joi.string().required().min(2), houseNumber: joi.string().required().min(1), zipcode: joi.string().min(2), lat: joi.number(), lng: joi.number() }), owner: joi.string().min(2)
@@ -65,6 +65,7 @@ router.post("/", auth, async (req, res) => {
         //Add the new card and save
         card = new Card(req.body);
         card.save();
+
         //return response
         res.status(201).send(`Card "${card.title}" was added successfully.`);
 
@@ -77,18 +78,19 @@ router.post("/", auth, async (req, res) => {
 router.put("/:_id", auth, async (req, res) => {
 
     try {
+        // Confirm if Owner or Admin
         if (req.payload.role != "admin" && req.payload.email != req.body.owner)
             return res.status(400).send("Only Admin or card owner are alloud to update this card")
 
-        //1. joi validation
+        // joi validation
         const { error } = cardSchema.validate(req.body);
         if (error) return res.status(400).send(error);
 
-        //2. Verify&Update user by req _id
+        // Verify&Update user by req _id
         const card = await Card.findOneAndUpdate({ _id: req.params._id }, req.body, { new: true });
         if (!card) return res.status(400).send("No such card")
 
-        //3. return response
+        // return response
         res.status(200).send(`${card.title} was updated successfully!!`)
 
     } catch (error) {
@@ -97,16 +99,21 @@ router.put("/:_id", auth, async (req, res) => {
 })
 
 //Delete card
-router.delete("/:_id", async (req, res) => {
+router.delete("/:_id", auth, async (req, res) => {
     try {
+        // Confirm if Owner or Admin
+        if (req.payload.role != "admin" && req.payload.email != req.body.owner)
+            return res.status(400).send("Only Admin or card owner are alloud to delete this card")
+
+        // Check if exist and delete Card
         const card = await Card.findOneAndDelete({ _id: req.params._id });
         if (!card) return res.status(400).send("This Business Card details are not available...");
-        res.status(200).send(card);
+
+        // return response
+        res.status(200).send(`${card.title} was deleted successfully!!`);
     } catch (error) {
         res.status(400).send(error);
     }
 })
-
-
 
 module.exports = router;
